feat: enable the suspicious lint category

+++ test/stroke-buffer.test.ts
@@ -143,7 +143,7 @@ class FakeDoc {
 		this.layer = {
 			width: W,
 			height: H,
-			data: (snap.data.data as Uint8ClampedArray).slice(),
+			data: snap.data.data.slice(),
 		};
 	}
